fix(scan): tolerate three real-world QR shapes that broke parsing
Two real receipts the user couldn't scan exposed three independent issues:

1. **Padded encoding flag (魯魯札札 cafe)** — the right-code's trailing space
   padding bleeds into the left QR's encoding flag, arriving as '0  '. The
   strict `when (tail[2]) { "0" -> BIG5; "1" -> UTF8; else -> throw }` ate
   the whole scan. Trim before matching.

2. **Header-only left QR (福懋加油站 gas station)** — gas-station POS emits
   a truncated left code: fixed prefix + AES verification only, NO `*` 自定區
   and NO :品目筆數: ... detail section. All items live in the right code.
   Old behaviour: throw 'missing 營業人自定區'. New behaviour: accept it as
   a valid header-only invoice with empty left items + default UTF-8
   encoding. Dropped bad-03 from malformed fixtures (it tested for exactly
   this case, which is actually legal).

3. **Decimal qty / price (gas litres)** — 95Plus 無鉛: 30.32 L × NT$33.9.
   `toIntOrNull()` returned null and groupItems silently dropped the row.
   Changed ParsedItem.quantity Int → Double; both qty and price now parse
   through `toDoubleOrNull`; unitPrice rounds to Int (the DB schema's
   contract — sub-NT$1 precision is irrelevant for receipt items, the
   header totalAmount is the source of truth for money totals). amount =
   (qty * unitPrice).roundToInt().

Tests: parser fixture comparator now coerces expected.quantity to Double;
the two inline tests (麥味登 right-code padding) updated to 1.0; three new
tests lock in (1) padded flag, (2) header-only left, (3) decimal qty/price.

+++ android/feature/scan/src/main/kotlin/tw/invoicewallet/feature/scan/qr/EInvoiceQrModel.kt
@@ -5,8 +5,10 @@ import kotlinx.datetime.LocalDate
 /** Text encoding declared by the left QR's flag (0 = Big5, 1 = UTF-8). */
 enum class InvoiceTextEncoding { UTF8, BIG5 }
 
-/** One line item decoded from a QR code. */
-data class ParsedItem(val name: String, val quantity: Int, val unitPrice: Int)
+/** One line item decoded from a QR code. Quantity is [Double] because gas-station
+ *  invoices sell by litres (e.g. 30.32 L × NT$33.9). Unit price stays Int — receipt
+ *  items round to the nearest dollar, the < NT$1 precision loss is acceptable. */
+data class ParsedItem(val name: String, val quantity: Double, val unitPrice: Int)
 
 /** The fixed + delimited fields decoded from the LEFT QR code. */
 data class LeftCode(
